requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.0.0
numpy>=1.21.0
scipy>=1.7.0
pyyaml>=6.0
matplotlib>=3.5.0

# Training libraries
unsloth
trl>=0.7.0
accelerate>=0.20.0
peft>=0.4.0

# Optional dependencies
wandb>=0.15.0
ipywidgets>=8.0.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0